Levels work!
now if you are to die in the game and then you go back a level. Also you are able to go between and progress to the next level if you hit the tile with the tag "Win".  One feature that i would like to add is for the code to allow the death and the victory (not made yet) animations play before you switch scenes.